Renamed option item class file name
