+++ src/components/editor-input/mention-plugin/mention-option-item.js
